Body is displayed and meta info is hidden
Added a 'More Info' button that expands the list of meta data that used
to be displayed.

Body field is now displayed under the search result

frag size for highlighted fields is now set to 250 (controllers/catalog)

TODO: The implementation of the body being displayed in
views/catalog/_index_default is very poor.  Should be reworked at some
point