Bar Graph functionality added
AJAX call is working
    -However, all the functionality follows up in the anonymous
function that is called upon success.  This should be able to return
the json object instead of need to call this locally.

Added TODO.txt to outline and prioritize necessary changes

added a place to render the landing page information from solr once it
is implemented