Commenting Bookmarks and Scrubbing body display
The bookmarks of the document header in the catalog were commented out
(and the header drop down option to view bookmarks) due to the log in
functionality being take out.  The feature does not exist without log
in.

Also, the body of the document in the catalog now uses the highlighted
field "body" instead of the solr field "body".

This change was necessary because "body" was not a default field in
LWS, but it is a default highlighted field.